+++ UVtools.GUI/Properties/Resources.Designer.cs
@@ -553,6 +553,16 @@ namespace UVtools.GUI.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap microchip_16x16 {
+            get {
+                object obj = ResourceManager.GetObject("microchip_16x16", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap move_16x16 {
             get {
                 object obj = ResourceManager.GetObject("move-16x16", resourceCulture);